ACROSS-monitoring-stack/Kubernetes/docker/ml/requirements.txt
requests
kafka-python
sympy
